Rewrite createRegistry.cpp.
It now accepts --plugin-directory and --registry-file arguments so
that it can read plugins and write registries outside the installed
prefix, and thereby not require privileges to run.

Minor difference: the registry.txt now says `libfoo.la' rather than
`libfoo' as it did before.

I was surprised to observe that this seemed to work with DESTDIR
installations, until I remembered that I had already installed the
libraries on which the plugins depended.  Oops.  (Naturally when you
try to dlopen $DESTDIR$PREFIX/lib/proto/foo.so and it depends on
$PREFIX/lib/libbar.so, it won't work unless you've already installed
to $PREFIX.)  So this was probably a waste of time; we really need a
different mechanism for the plugin registry altogether.

git-svn-id: file:///Users/jakebeal/projects/_fin/proto/svn/proto/trunk@461 4c76e1c1-4b85-4c65-98a1-c45dc8243d55