Convert C-style (Foo *)x to &dynamic_cast<Foo &>(*x) in neocompiler.
Use static_cast for non-class casts.

On a hunt for memory corruption and a null pointer somewhere it
shouldn't be.  Haven't found it yet, but this make the code more
robust anyway by detecting bad casts before their results are used.

Many C-style casts still remain in the paleocompiler and simulator.

git-svn-id: file:///Users/jakebeal/projects/_fin/proto/svn/proto/trunk@488 4c76e1c1-4b85-4c65-98a1-c45dc8243d55